Allow elevation to be negative
git-svn-id: http://feedvalidator.googlecode.com/svn/trunk@745 34a10a33-d82d-0410-ba40-81c901463ecc

+++ src/feedvalidator/extension.py
@@ -633,7 +633,7 @@ class extension_item(extension_channel_item):
     return text()
 
   def do_georss_elev(self):
-    return Float()
+    return decimal()
 
   def do_georss_floor(self):
     return Integer()
